backend/requirements.txt
django-cors-headers
djangorestframework
python-dotenv
djangorestframework-simplejwt

Django>=4.2
requests
PyPDF2
pdf2image
pytesseract
Pillow


Django>=5.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.1
python-dotenv>=1.0.1

# AI / PDF tooling
requests>=2.32.3
pdfminer.six>=20231228
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.3.0
